Done with the custom listview and tested

+++ app/src/main/java/us/bojie/earthquakebo/model/Earthquake.java
@@ -6,21 +6,21 @@ import java.util.Date;
  * Created by bjiang on 8/4/16.
  */
 public class Earthquake {
-    private float mMagnitude;
+    private double mMagnitude;
     private String mLocation;
     private Date mTime;
 
-    public Earthquake(float mMagnitude, String mLocation, Date mTime) {
+    public Earthquake(double mMagnitude, String mLocation, Date mTime) {
         this.mMagnitude = mMagnitude;
         this.mLocation = mLocation;
         this.mTime = mTime;
     }
 
-    public float getmMagnitude() {
+    public double getmMagnitude() {
         return mMagnitude;
     }
 
-    public void setmMagnitude(float mMagnitude) {
+    public void setmMagnitude(double mMagnitude) {
         this.mMagnitude = mMagnitude;
     }
